input: xpad: Prevent corruption of urb request.
The xpad_send_led_command() and xpad_play_effect() functions submit urb
requests, but do not
wait for the previous urb request to complete before using the same
resources to
submit a new request.  This can lead to unpredictable undesirable
effects, including
memory corruption and dereferencing of NULL pointers(and needless to
say, kernel
panics).

Fix the issue by introducing a busy flag set on submission of the urb
request, and
cleared on urb request completion.  If this flag is set while in
xpad_send_led_command()
or xpad_play_effect(), the led/rumble packet data is buffered, and will
be sent from
the urb request completion routine when the current urb request
finishes.

Patch tested with rumble against a Logitech F510 and an X-Box v1 pad.