Input: xpad: update Xbox One Force Feedback Support
Add force feedback support for Xbox One controllers.
There's apparently a serial number woven into both input and output
packets;
neglecting to specify a valid serial number causes the controller to
ignore
the rumble packets.

The scale of the rumble was also apparently halved in the packets.

The initialization packet had to be changed to allow force feedback to
work.