Input: xpad: query Wireless controller state at init
We initializing the driver/device, we really don't know how many
controllers are connected.  So send a "query presence" command to the
base-station. (Command discovered by Zachary Lund)

Note, this means we now do not "preallocate" all 4 devices when a single
wireless base station is seen, but require the device to be properly
connected to the base station before that can happen.  The allocation of
the device happens in the next patch, not here, so in a way, this patch
breaks all wireless devices...

based on patch by 
"Pierre-Loup A. Griffais" <pgriffais@valvesoftware.com>

presence packet taken from:
https://github.com/computerquip/xpad5