Input: xpad: remove bulk out URB
this code is redundand with xpad_send_led_command which does the same
thing using irq_out.
It is also misnamed as it is likely to be an irq out and not a bulk out. 
Furthermore it was probably wrong from the start as setting the LEDs did
not work on my x360w controller before fixing xpad_send_led_command in
the last patch series.