Input: xpad: do not submit active URBs
track the active status of the irq_out URB to prevent submission while
it is active. Failure to do so results in the "URB submitted while
active" warning/ stacktrace.

Also add missing locking around xpad->odata usages: as we can not use
mutexes in xpad_play_effect, replace the mutex by a spinlock.