Support all Xbox Elite 2 packet types
After doing some research, it seems there have historically been 3
different packet formats used with the Elite 2 controller. These formats
differ in how they deliver certain newer inputs like the paddles, so we
have to detect and account for these when trying to pull paddle data.

It's POSSIBLE that there is also a 4th packet format that existed for a
brief period in early firmware versions, but it is likely that this
fourth type only appeared on bluetooth connections.