Support KEY_RECORD for 8BitDo Pro 2...for Xbox
This controller is a "Series S|X": it has the "share" button below the
big Xbox button. However, it reports the status of that button in a
different offset than the official controller.

The button is still recognized by official Microsoft driver in Windows,
so it looks like that both offsets are used in the wild.

Signed-off-by: Leonardo Brondani Schenkel <leonardo@schenkel.net>