Input: xpad - add support for Xbox ONE liquid metal controller
I found that my controller wasn't working with this driver.  The LED would light up and it could rumble but was not registering any button presses or joystick movements.  I suspected the problem might be that it was not being sent some required init packet so I tried sending the packets being used in xboxone_init_packets and found that sending xboxone_pdp_auth got the controller working.  At least for my controller xboxone_pdp_led_on was not strictly necessary but I decided to follow the same pattern used by the other two controllers.