Input: xpad - added unofficial Xbox 360 wireless receiver clone
I've added the DeviceID of a very popular Xbox 360 Wireless Receiver.

Although it mimics the Microsoft's VendorID, it is in fact a clone. Taking into account that the original Microsoft Receiver is not being manufactured anymore, this drive can solve dpad issues encontered by those who still use the original 360 Wireless controller but are using a receiver clone.

Signed-off-by: Nilton Perim Neto niltonperimneto@gmail.com