Fixes for the XBOX 360 usb control messages
1)
XBOX 360 usb control messages for an interface, need the correct Interface specified.
Currently this is set to a fixed value of Interface 0.
This is needed for usb devices that do not have the XBOX 360 on the first Interface, and/or for usb devices that have multiple XBOX 360 Interfaces.

2)
Removed the cpu_to_le16() usage from the recently added 'QUIRK_360_START' packets, as this is handled by the function usb_control_msg() and not by the user of this function.

Signed-off-by: M Bakker <mbakker@embenco.nl>